+++ API/src/main/java/fr/maxlego08/essentials/api/EssentialsPlugin.java
@@ -362,4 +362,6 @@ public interface EssentialsPlugin extends Plugin {
      * @return an {@link Optional} containing the instance if successful, or empty if not
      */
     <T> Optional<T> createInstance(String className, boolean displayLog);
+
+    long getServerStartupTime();
 }