+++ src/main/java/fr/maxlego08/essentials/commands/commands/utils/admins/CommandGod.java
@@ -24,7 +24,7 @@ public class CommandGod extends VCommand {
 
         Player player = this.argAsPlayer(0, this.player);
 
-        if (this.player == null) {
+        if (player == null) {
             return CommandResultType.SYNTAX_ERROR;
         }