+++ src/main/java/fr/maxlego08/essentials/commands/commands/warp/CommandWarps.java
@@ -22,8 +22,15 @@ public class CommandWarps extends VCommand {
     @Override
     protected CommandResultType perform(EssentialsPlugin plugin) {
 
+        WarpModule warpModule = plugin.getModuleManager().getModule(WarpModule.class);
+        if (warpModule.isEnableInventory()) {
+            plugin.getInventoryManager().openInventory(this.player, plugin, "warps");
+            return CommandResultType.DEFAULT;
+        }
+
         List<String> warps = plugin.getWarps().stream().filter(warp -> warp.hasPermission(sender)).map(warp -> getMessage(Message.COMMAND_WARP_LIST_INFO, "%name%", warp.name())).toList();
         message(sender, Message.COMMAND_WARP_LIST, "%destinations%", Strings.join(warps, ','));
+
         return CommandResultType.SUCCESS;
     }
 }